repo: move USBProxyFilter to filters/
