+++ facedancer/filters/__init__.py
@@ -1,0 +1,3 @@
+from .base     import  USBProxyFilter
+from .logging  import  USBProxyPrettyPrintFilter
+from .standard import  USBProxySetupFilters
